Takes Now Store their own Directory Name
Does what it says in the description.

Why did I do this? In order to load all the frames, the take needs to open and query the folder containing such frames.

Without storing the take folder name as part of the take, how will projects be opened? In theory, we could re-constructs the take Directory Name from the shotNumber and takeNumber.
With such a system, ANY TIME we change makeTakeDirectoryName In ANY MANNER WHATSOEVER, It would break all future loading.  This is an odious onus on the dev team.

Thus Why I want the take directory name to be stored as part of the take. If I save a project, install an update in which the take directory name construction is changed, and try to re-open the project, it wouldn't get bamboozled by the name change