ProjectFilesContextProvider now loads from file
Added a new method: loadProjectFromDisk. It consistently was thwarted by the "Schroedingers directory: This directory exists, but is not in the database". The Database stores the folder "Gee2Cameras!.boatsfiles" as "Gee2Cameras!", which hampered it's ability to be queried without a cursed conversion function between the friendly name and the boatsfiles name.